requirements.txt
dj-database-url==0.4.1
Django==1.11.1
django-angular==1.1.3
django-filter==1.0.4
djangorestframework==3.7.0
gunicorn==19.6.0
Markdown==2.6.9
psycopg2==2.6.2
pytz==2017.2
whitenoise==3.2
